Bare minimum needed to add navbar to all pages and keep the pages running
Turns out, to get it working, there is hardly any porting job to do. Maintaining a consistent style is another, but given the time we have it's been put on the backburner. If it work, it works.

Also added link to the big portrait that points back to the main page.